test: feature cache

+++ tests/fixtures.py
@@ -4,12 +4,15 @@ import pathlib
 import shutil
 import subprocess
 import tempfile
-from typing import Iterator
+from typing import Callable, Iterator, Optional, Tuple, cast
 
 from dvc.repo import Repo as DvcRepo
 from pytest_cases import fixture
 
 from kedro_dvc.create_sample_project import create_sample_project
+
+CACHE_DIR = pathlib.Path(__file__).parent / ".fixture-cache"
+APP_DIR = pathlib.Path(__file__).parent.parent
 
 
 @contextlib.contextmanager
@@ -46,17 +49,51 @@ def fix_tmp_dir_session() -> Iterator[pathlib.Path]:
         yield dir
 
 
+@contextlib.contextmanager
+def to_memoized_dir(
+    cache_dir: pathlib.Path,
+) -> Iterator[Tuple[pathlib.Path, Optional[Callable[[], None]]]]:
+    """
+    Create temp dir; fill from cache if exists; yield context within it.
+
+    If the cache dir is empty, `save_cache` will be a function that
+    will save the cache. Otherwise, it will be None.
+
+    Returns: tuple of (`dir`, `save_cache`):
+        dir: pathlib.Path to the created directory
+        save_cache: optional callable to save the cache to the `cache_dir`
+    """
+    with to_tmp_dir() as tmp_dir:
+        save_cache: Optional[Callable[[], None]] = None
+        breakpoint()
+        if cache_dir.exists():
+            shutil.copytree(str(cache_dir), str(tmp_dir))
+            save_cache = cast(
+                Callable[[], None],
+                lambda: shutil.copytree(
+                    str(tmp_dir), str(cache_dir), dirs_exist_ok=True
+                ),
+            )
+        yield tmp_dir, save_cache
+
+
 @fixture(name="dvc_repo_session", scope="session")  # type: ignore
 def fix_dvc_repo_session(tmp_dir_session: pathlib.Path) -> Iterator[DvcRepo]:
     """
     Create dvc repo; test cwd will be within repo dir.
     """
-    subprocess.check_call(["git", "init"])
-    dvc = DvcRepo.init(".", subdir=True)
-    try:
-        yield dvc
-    finally:
-        dvc.close()
+    cache_dir = CACHE_DIR / "dvc_repo_session"
+    with to_memoized_dir(cache_dir) as (dir, save_cache):
+        if save_cache:
+            subprocess.check_call(["git", "init"])
+            dvc = DvcRepo.init(".", subdir=True)
+            save_cache()
+        else:
+            dvc = DvcRepo(root_dir=dir)
+        try:
+            yield dvc
+        finally:
+            dvc.close()
 
 
 @fixture(name="dvc_repo")  # type: ignore
@@ -65,8 +102,8 @@ def fix_dvc_repo(dvc_repo_session: DvcRepo) -> Iterator[DvcRepo]:
     Create dvc repo (copying session repo); cwd within repo.
     """
     with to_tmp_dir() as dir:
+        shutil.copytree(dvc_repo_session.root_dir, dir)
         dvc_repo = DvcRepo(root_dir=dir)
-        shutil.copytree(dvc_repo_session.root_dir, dvc_repo.root_dir)
         try:
             yield dvc_repo
         finally:
@@ -80,10 +117,13 @@ def fix_empty_kedro_repo_session() -> Iterator[pathlib.Path]:
 
     Session scoped.
     """
-    with to_tmp_dir() as dir:
-        create_sample_project("test")
-        shutil.move("tmp/test", ".")
-        shutil.rmtree("tmp")
+    cache_dir = CACHE_DIR / "empty_kedro_repo_session"
+    with to_memoized_dir(cache_dir) as (dir, save_cache):
+        if save_cache:
+            create_sample_project("test", kedro_dvc_path=APP_DIR)
+            shutil.move("tmp/test", ".")
+            shutil.rmtree("tmp")
+            save_cache()
         yield dir
 
 
@@ -96,13 +136,18 @@ def fix_empty_repo_session(
 
     Session scoped.
     """
-    with to_tmp_dir() as dir:
-        shutil.copytree(dvc_repo_session.root_dir, dir)
-        shutil.copytree(empty_kedro_repo_session, dir)
-        empty_repo = DvcRepo(root_dir=dir)
-        empty_repo.git.add_commit(".", "feat: first commit of empty repo")
+    cache_dir = CACHE_DIR / "empty_repo_session"
+    with to_memoized_dir(cache_dir) as (dir, save_cache):
+        if save_cache:
+            shutil.copytree(dvc_repo_session.root_dir, dir)
+            shutil.copytree(empty_kedro_repo_session, dir)
+            empty_repo = DvcRepo(root_dir=dir)
+            empty_repo.git.add_commit(".", "feat: first commit of empty repo")
+            save_cache()
+        else:
+            empty_repo = DvcRepo(root_dir=dir)
         try:
-            yield dir
+            yield empty_repo
         finally:
             empty_repo.close()
 
@@ -119,3 +164,11 @@ def fix_empty_repo(empty_repo_session: DvcRepo) -> Iterator[DvcRepo]:
             yield dvc_repo
         finally:
             dvc_repo.close()
+
+
+def clear_fixture_cache() -> None:
+    """
+    Removes the cached fixtures.
+    """
+    if CACHE_DIR.exists():
+        shutil.rmtree(CACHE_DIR)
